+++ pdf-stats-service/src/main/java/com/github/zukkari/pdfstatsservice/stats/service/RenderCountService.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Service
 public class RenderCountService {
@@ -21,14 +21,14 @@ public class RenderCountService {
     }
 
     public void increment() {
-        RenderCount stats = repository.findById(1L).map(RenderCount::increment).orElseGet(() -> RenderCount.of(BigDecimal.ONE));
+        RenderCount stats = repository.findById(1L).map(RenderCount::increment).orElseGet(() -> RenderCount.of(BigInteger.ONE));
 
         repository.save(stats);
         log.info("New pdf generation count '{}'", stats.getCount());
     }
 
     @Cacheable(cacheNames = "count")
-    public BigDecimal getCount() {
-        return repository.findById(1L).map(RenderCount::getCount).orElse(BigDecimal.ZERO);
+    public BigInteger getCount() {
+        return repository.findById(1L).map(RenderCount::getCount).orElse(BigInteger.ZERO);
     }
 }
